dashboard-v2/requirements.txt
django>=5.1,<6.0
pyyaml>=6.0
